+++ src/main/java/org/cotato/csquiz/domain/recruitment/service/RecruitmentInformationService.java
@@ -7,8 +7,8 @@ import org.cotato.csquiz.common.error.ErrorCode;
 import org.cotato.csquiz.common.error.exception.AppException;
 import org.cotato.csquiz.domain.generation.embedded.Period;
 import org.cotato.csquiz.domain.recruitment.entity.RecruitmentInformation;
+import org.cotato.csquiz.domain.recruitment.scheduler.RecruitmentScheduler;
 import org.cotato.csquiz.domain.recruitment.service.component.RecruitmentInformationReader;
-import org.cotato.csquiz.domain.schedule.RecruitmentScheduler;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.support.TransactionSynchronization;
